cart provider 연동하고 영어로 안내문구 바꿈

+++ lib/pages/item_list.dart
@@ -33,7 +33,7 @@ class _ItemListState extends State<ItemList> {
   }
 
   /// 가격을 천 단위 콤마로 포맷해주는 함수
-  String _formatPrice(int price) {
+  String _formatPrice(double price) {
     return price.toString().replaceAllMapped(
           RegExp(r'(\d)(?=(\d{3})+(?!\d))'),
           (match) => '${match[1]},',
